Optimize LCP with Server Components and deferred analytics
Phase 1 - Config & Analytics:
- Add optimizePackageImports for lucide-react tree-shaking
- Change GTM from afterInteractive to lazyOnload strategy
- Defer web-vitals reporting with requestIdleCallback

Phase 2 - Component Optimization:
- Convert Hero to Server Component for faster LCP
- Extract CTA buttons to HeroActions client component
- Static content (h1, badges, text) now server-rendered

Expected impact:
- LCP: 3.7s → ~2.3-2.5s
- Performance: 87 → 92-95
- Render blocking: 340ms → ~100ms
