+++ src/components/sections/hero.tsx
@@ -1,11 +1,7 @@
-'use client';
-
-import Link from 'next/link';
-import { CheckCircle, Award, GraduationCap, Zap, Shield } from 'lucide-react';
-import { Button } from '@/components/ui/button';
+import { Award, GraduationCap, Zap, Shield } from 'lucide-react';
 import { Container, Section } from '@/components/layout';
 import { SITE_CONFIG } from '@/lib/constants';
-import { trackEnrollClick, trackScheduleCall, trackEvent } from '@/lib/analytics';
+import { HeroActions } from './hero-actions';
 
 const socialProofBadges = [
   {
@@ -28,18 +24,6 @@ const socialProofBadges = [
 ];
 
 export function Hero() {
-  const handleEnrollClick = () => {
-    trackEnrollClick('hero');
-  };
-
-  const handleScheduleClick = () => {
-    trackScheduleCall('hero');
-  };
-
-  const handlePhoneClick = () => {
-    trackEvent('phone_call', { location: 'hero' });
-  };
-
   return (
     <Section className="pt-8 md:pt-16 pb-16 md:pb-24" background="white">
       <Container>
@@ -86,33 +70,8 @@ export function Hero() {
             ))}
           </div>
 
-          {/* CTAs */}
-          <div className="flex flex-col sm:flex-row gap-4 justify-center mb-8">
-            <Button
-              size="lg"
-              className="bg-amber-500 hover:bg-amber-600 text-black font-semibold"
-              asChild
-              onClick={handleEnrollClick}
-            >
-              <a href={SITE_CONFIG.links.courseCatalog} target="_blank" rel="noopener noreferrer">
-                Start Your Training
-              </a>
-            </Button>
-            <Button size="lg" variant="outline" asChild onClick={handleScheduleClick}>
-              <a href={SITE_CONFIG.links.scheduleCall} target="_blank" rel="noopener noreferrer">
-                Book an Online Meeting
-              </a>
-            </Button>
-          </div>
-
-          {/* Phone CTA */}
-          <a
-            href={`tel:${SITE_CONFIG.phone}`}
-            onClick={handlePhoneClick}
-            className="inline-flex items-center text-gray-600 hover:text-primary transition-colors"
-          >
-            Call Now: {SITE_CONFIG.phone}
-          </a>
+          {/* CTAs - Client Component for tracking */}
+          <HeroActions />
         </div>
       </Container>
     </Section>
